fancy-cat: target xgene1 on arm64 Linux as closest to `-march=armv8-a`
X-Gene 1 is only CPU target that does not enable extra extensions like
crc. This does mean it is less optimal for the average user but our
official baseline target on ARM64 Linux is armv8-a. The CPU target
can be adjusted to cortex_a34 or cortex_a35 if we change our official
baseline to require `aes+crc+sha2`, which is available on pretty much
all other ARMv8 CPUs other than X-Gene 1.